backend/requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12,<4.0
pymongo>=3.11,<4.0
requests>=2.25,<3.0
python-dotenv>=0.15,<1.0
celery>=5.1,<6.0
pandas>=1.1,<2.0
PyPDF2>=1.26,<2.0
gunicorn>=20.1,<21.0